requirements.txt
denoiser==0.1.5
termcolor
datasets
datasets[audio]
numpy==1.23.5
easydict
tqdm
pytorch-crf
numba
librosa
PyYAML
scikit-learn==0.23.2
scipy==1.5.2
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
jiwer==3.0.3
evaluate
matplotlib
